+++ action_suite.h
@@ -10,7 +10,7 @@
 #ifndef ACTION_H
 #define ACTION_H
 
-void pick_phi(int i, int j, int (&phi)[3], double *** Lattice, bool old_lattice);
+void pick_phi(int i, int j, double (&phi)[3], double *** Lattice, bool old_lattice);
 void make_triangles(int i, int j, int len, int (&triangles)[8][3][2]);
 double QL_triangle(int current_triangle[3][2], double *** Lattice, bool arcsin, bool old_lattice);
 double Q_lattice(double *** Lattice, int len, bool old_lattice);
